+++ src/DotNet.Meteor.Debug/DotNet.Meteor.Debug.csproj
@@ -3,6 +3,7 @@
 
 	<PropertyGroup>
 		<Nullable>disable</Nullable>
+		<ImplicitUsings>disable</ImplicitUsings>
 	</PropertyGroup>
 
 	<ItemGroup>
