+++ src/DotNet.Meteor.Debug/Extensions/Handles.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace DotNet.Meteor.Debug.Utilities;
+namespace DotNet.Meteor.Debug.Extensions;
 
 public class Handles<T> {
     private const int START_HANDLE = 1000;
